+++ src/components/ShareModal.jsx
@@ -1,6 +1,6 @@
 import { Dialog, IconButton, Input } from "@mui/material";
-import React, { useState } from "react";
 
+import React from "react";
 import logo from "../assets/logo.png";
 import { useSnackbar } from "notistack";
