+++ src/pages/Login/index.jsx
@@ -55,7 +55,7 @@ const LoginScreen = () => {
             variant: "error",
           });
         } else {
-          enqueueSnackbar("Error Occured!", {
+          enqueueSnackbar(error.message, {
             variant: "error",
           });
         }
@@ -81,7 +81,7 @@ const LoginScreen = () => {
               variant: "error",
             });
           } else {
-            enqueueSnackbar("An error occcured!", {
+            enqueueSnackbar(error.message, {
               variant: "error",
             });
           }
@@ -104,7 +104,7 @@ const LoginScreen = () => {
             variant: "error",
           });
         } else {
-          enqueueSnackbar("An error occcured!", {
+          enqueueSnackbar(error.message, {
             variant: "error",
           });
         }
@@ -153,7 +153,6 @@ const LoginScreen = () => {
               backgroundColor: "white",
               boxShadow: "0 0 20px rgba(104, 85, 224, 0.2)",
               borderRadius: "4px",
-              // padding: "10px",
             }}
           >
             <input
